WIP: decompose FolioReaderCenter, try to fix resizing
