+++ Cartfile
@@ -1,4 +1,5 @@
 github "ZipArchive/ZipArchive" ~> 2.1.1
+github "weichsel/ZIPFoundation" == 0.9.11
 github "cxa/MenuItemKit" == 3.1.3
 github "alexpopov/ZFDragableModalTransition" "merge-carthage-into-zoonooz"
 github "tadija/AEXML" == 4.3.3
