+++ Cartfile.resolved
@@ -4,3 +4,4 @@ github "alexpopov/ZFDragableModalTransition" "8da951efb4202385630d1cf7104b60d120
 github "cxa/MenuItemKit" "3.1.3"
 github "realm/realm-cocoa" "v5.5.2"
 github "tadija/AEXML" "4.3.3"
+github "weichsel/ZIPFoundation" "0.9.11"
