+++ Sources/FolioReaderKit/Center/CenterViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  CenterViewModel.swift
+//  FolioReaderKit
+//
+//  Created by 京太郎 on 2021/9/14.
+//  Copyright © 2021 FolioReader. All rights reserved.
+//
+
+import Foundation
+
+class FolioReaderCenterViewModel {
+    
+}
+
