+++ Sources/FolioReaderKit/Center/ChapterListDelegation.swift
@@ -1,0 +1,44 @@
+//
+//  ChapterListDelegation.swift
+//  FolioReaderKit
+//
+//  Created by 京太郎 on 2021/9/14.
+//  Copyright © 2021 FolioReader. All rights reserved.
+//
+
+import Foundation
+
+extension FolioReaderCenter: FolioReaderChapterListDelegate {
+    
+    func chapterList(_ chapterList: FolioReaderChapterList, didSelectRowAtIndexPath indexPath: IndexPath, withTocReference reference: FRTocReference) {
+        if readerConfig.debug.contains(.functionTrace) { folioLogger("ENTER") }
+
+        let item = findPageByResource(reference)
+        
+        if item < totalPages {
+            let indexPath = IndexPath(row: item, section: 0)
+            changePageWith(indexPath: indexPath, animated: false, completion: { () -> Void in
+                self.updateCurrentPage()
+            })
+            tempReference = reference
+        } else {
+            print("Failed to load book because the requested resource is missing.")
+        }
+    }
+    
+    func chapterList(didDismissedChapterList chapterList: FolioReaderChapterList) {
+        updateCurrentPage()
+        if readerConfig.debug.contains(.functionTrace) { folioLogger("ENTER") }
+
+        // Move to #fragment
+        if let reference = tempReference {
+            if let fragmentID = reference.fragmentID, let currentPage = currentPage , fragmentID != "" {
+                currentPage.handleAnchor(reference.fragmentID!, avoidBeginningAnchors: true, animated: true)
+            }
+            tempReference = nil
+        }
+    }
+    
+    
+}
+
