+++ Sources/FolioReaderKit/Center/PageDelegation.swift
@@ -1,0 +1,71 @@
+//
+//  PageDelegation.swift
+//  FolioReaderKit
+//
+//  Created by 京太郎 on 2021/9/14.
+//  Copyright © 2021 FolioReader. All rights reserved.
+//
+
+import Foundation
+
+extension FolioReaderCenter: FolioReaderPageDelegate {
+
+    public func pageDidLoad(_ page: FolioReaderPage) {
+        if readerConfig.debug.contains(.functionTrace) { folioLogger("ENTER") }
+
+        if self.readerConfig.loadSavedPositionForCurrentBook, let position = folioReader.savedPositionForCurrentBook {
+//        if self.readerConfig.loadSavedPositionForCurrentBook, let position = self.readerConfig.savedPositionForCurrentBook {
+//            folioReader.savedPositionForCurrentBook = position
+            let pageNumber = position["pageNumber"] as? Int
+            let offset = self.readerConfig.isDirection(position["pageOffsetY"], position["pageOffsetX"], position["pageOffsetY"]) as? CGFloat
+            let pageOffset = offset
+
+            if isFirstLoad {
+                updateCurrentPage(page)
+                isFirstLoad = false
+
+                if (self.currentPageNumber == pageNumber && pageOffset > 0) {
+                    page.scrollPageToOffset(pageOffset!, animated: false)
+                }
+                
+            } else if (self.isScrolling == false && folioReader.needsRTLChange == true) {
+                page.scrollPageToBottom()
+            }
+        } else if isFirstLoad {
+            updateCurrentPage(page)
+            isFirstLoad = false
+        }
+
+        updateCurrentPage(page)
+        page.webView?.isHidden = false
+        
+        // Go to fragment if needed
+        if let fragmentID = tempFragment, let currentPage = currentPage , fragmentID != "" {
+            currentPage.handleAnchor(fragmentID, avoidBeginningAnchors: true, animated: true)
+            tempFragment = nil
+        }
+        
+        if (readerConfig.scrollDirection == .horizontalWithVerticalContent),
+            let offsetPoint = self.currentWebViewScrollPositions[page.pageNumber - 1] {
+            page.webView?.scrollView.setContentOffset(offsetPoint, animated: false)
+        }
+        
+        // Pass the event to the centers `pageDelegate`
+        pageDelegate?.pageDidLoad?(page)
+    }
+    
+    public func pageWillLoad(_ page: FolioReaderPage) {
+        if readerConfig.debug.contains(.functionTrace) { folioLogger("ENTER") }
+
+        // Pass the event to the centers `pageDelegate`
+        pageDelegate?.pageWillLoad?(page)
+    }
+    
+    public func pageTap(_ recognizer: UITapGestureRecognizer) {
+        if readerConfig.debug.contains(.functionTrace) { folioLogger("ENTER") }
+
+        // Pass the event to the centers `pageDelegate`
+        pageDelegate?.pageTap?(recognizer)
+    }
+    
+}
